resturcuting of e2e tests
